+++ sdk/communication/azure-communication-identity/src/samples/java/com/azure/communication/identity/ReadmeSamples.java
@@ -119,7 +119,7 @@ public class ReadmeSamples {
         List<CommunicationTokenScope> scopes = 
             new ArrayList<>(Arrays.asList(CommunicationTokenScope.CHAT));
 
-        AccessToken userToken = communicationIdentityClient.issueToken(user, scopes);
+        AccessToken userToken = communicationIdentityClient.getToken(user, scopes);
         System.out.println("User token value: " + userToken.getToken());
         System.out.println("Expires at: " + userToken.getExpiresAt());
         return userToken;
@@ -133,7 +133,7 @@ public class ReadmeSamples {
         CommunicationUserIdentifier user = createNewUser();
         // Define a list of communication token scopes
         List<CommunicationTokenScope> scopes = Arrays.asList(CommunicationTokenScope.CHAT);
-        communicationIdentityClient.issueToken(user, scopes);
+        communicationIdentityClient.getToken(user, scopes);
         // revoke tokens issued for the specified user
         communicationIdentityClient.revokeTokens(user);
     }