+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/fluent/models/DomainInner.java
@@ -9,10 +9,12 @@ import com.azure.core.annotation.JsonFlatten;
 import com.azure.core.management.Resource;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.eventgrid.models.DomainProvisioningState;
+import com.azure.resourcemanager.eventgrid.models.IdentityInfo;
 import com.azure.resourcemanager.eventgrid.models.InboundIpRule;
 import com.azure.resourcemanager.eventgrid.models.InputSchema;
 import com.azure.resourcemanager.eventgrid.models.InputSchemaMapping;
 import com.azure.resourcemanager.eventgrid.models.PublicNetworkAccess;
+import com.azure.resourcemanager.eventgrid.models.ResourceSku;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
@@ -23,6 +25,18 @@ import java.util.Map;
 @Fluent
 public class DomainInner extends Resource {
     @JsonIgnore private final ClientLogger logger = new ClientLogger(DomainInner.class);
+
+    /*
+     * The Sku pricing tier for the domain.
+     */
+    @JsonProperty(value = "sku")
+    private ResourceSku sku;
+
+    /*
+     * Identity information for the resource.
+     */
+    @JsonProperty(value = "identity")
+    private IdentityInfo identity;
 
     /*
      * List of private endpoint connections.
@@ -78,6 +92,46 @@ public class DomainInner extends Resource {
      */
     @JsonProperty(value = "properties.inboundIpRules")
     private List<InboundIpRule> inboundIpRules;
+
+    /**
+     * Get the sku property: The Sku pricing tier for the domain.
+     *
+     * @return the sku value.
+     */
+    public ResourceSku sku() {
+        return this.sku;
+    }
+
+    /**
+     * Set the sku property: The Sku pricing tier for the domain.
+     *
+     * @param sku the sku value to set.
+     * @return the DomainInner object itself.
+     */
+    public DomainInner withSku(ResourceSku sku) {
+        this.sku = sku;
+        return this;
+    }
+
+    /**
+     * Get the identity property: Identity information for the resource.
+     *
+     * @return the identity value.
+     */
+    public IdentityInfo identity() {
+        return this.identity;
+    }
+
+    /**
+     * Set the identity property: Identity information for the resource.
+     *
+     * @param identity the identity value to set.
+     * @return the DomainInner object itself.
+     */
+    public DomainInner withIdentity(IdentityInfo identity) {
+        this.identity = identity;
+        return this;
+    }
 
     /**
      * Get the privateEndpointConnections property: List of private endpoint connections.
@@ -236,6 +290,12 @@ public class DomainInner extends Resource {
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
+        if (sku() != null) {
+            sku().validate();
+        }
+        if (identity() != null) {
+            identity().validate();
+        }
         if (privateEndpointConnections() != null) {
             privateEndpointConnections().forEach(e -> e.validate());
         }