+++ sdk/communication/azure-communication-identity/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## 1.0.0-beta.5 (Unreleased)
+### Breaking
+- `issueToken` is renamed to `getToken`.
 
 ## 1.0.0-beta.4 (2021-02-09)
 ### Breaking