+++ sdk/communication/azure-communication-identity/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-beta.5 (Unreleased)
 ### Breaking
 - `issueToken` is renamed to `getToken`.
+- `issueTokenWithResponse` is renamed to `getTokenWithReponse`.
 
 ## 1.0.0-beta.4 (2021-02-09)
 ### Breaking